preserve file history during rename
