+++ common/v2/config/index.ts
@@ -12,7 +12,7 @@ export {
 } from './constants';
 export { NODES_CONFIG } from './nodes';
 export { ROUTE_PATHS } from './routePaths';
-export { WALLET_INFO } from './walletInfo';
+export { WALLET_INFO } from './wallets';
 export { NETWORKS_CONFIG } from './networks';
 export { ETHSCAN_NETWORKS } from './ethScan';
 export { Fiats, AssetsData, ContractsData } from './cacheData';
